+++ kits/validationkit/tests/acceptance/src/ValidateDateTimeTest.php
@@ -683,6 +683,41 @@ class ValidateDateTimeTest extends TestCase
 
     }
 
+    #[TestDox('nullish() allows null')]
+    public function test_nullish_allows_null(): void
+    {
+        // ----------------------------------------------------------------
+        // explain your test
+
+        // ----------------------------------------------------------------
+        // shorthand
+
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $unit = Validate::nullish(Validate::dateTime());
+
+        // ----------------------------------------------------------------
+        // mock out any integrations
+
+        // ----------------------------------------------------------------
+        // pre-test checks
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = $unit->parse(null);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertNull($actualResult);
+
+        // ----------------------------------------------------------------
+        // clean up the database
+
+    }
+
     #[TestDox('default() provides fallback for null')]
     public function test_default_provides_fallback(): void
     {