+++ kits/validationkit/tests/acceptance/src/ValidateIntersectionTest.php
@@ -521,6 +521,48 @@ class ValidateIntersectionTest extends TestCase
 
     }
 
+    #[TestDox('nullish() allows null')]
+    public function test_nullish_allows_null(): void
+    {
+        // ----------------------------------------------------------------
+        // explain your test
+
+        // ----------------------------------------------------------------
+        // shorthand
+
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $unit = Validate::nullish(Validate::intersection(
+            left: Validate::object([
+                'name' => Validate::string(),
+            ]),
+            right: Validate::object([
+                'age' => Validate::int(),
+            ]),
+        ));
+
+        // ----------------------------------------------------------------
+        // mock out any integrations
+
+        // ----------------------------------------------------------------
+        // pre-test checks
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = $unit->parse(null);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertNull($actualResult);
+
+        // ----------------------------------------------------------------
+        // clean up the database
+
+    }
+
     #[TestDox('default() provides fallback for null')]
     public function test_default_provides_fallback(): void
     {